Progress_bar.py and enumerate addition.
