+++ tsfel/utils/__init__.py
@@ -2,3 +2,4 @@ from tsfel.utils.gSheetsFilters import *
 from tsfel.utils.calculate_complexity import *
 from tsfel.utils.signal_processing import *
 from tsfel.utils.add_personal_features import *
+from tsfel.utils.progress_bar import *
